feat: add condition to carditem to see cardmenu
