+++ components/CardMenu.jsx
@@ -19,38 +19,11 @@ export default function CardMenu(props) {
   const [editModalIsOpen, setEditModalIsOpen] = React.useState(false);
   const [error, setError] = useState(null);
   const [success, setSuccess] = useState(false);
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-  const router = useRouter();
-  const auth = getAuth(app);
-  const { enqueueSnackbar } = useSnackbar();
   const open = Boolean(anchorEl);
 
-  useEffect(() => {
-    // Check user's login status initially
-    const initialUser = auth.currentUser;
-    setIsLoggedIn(!!initialUser); // Set isLoggedIn based on the initial user
-
-    const unsubscribe = onAuthStateChanged(auth, (user) => {
-      if (user) {
-        setIsLoggedIn(true);
-      } else {
-        setIsLoggedIn(false);
-      }
-    });
-
-    // Clean up the subscription when component unmounts
-    return () => unsubscribe();
-  }, []);
-
   const handleEditClick = () => {
-    if (!isLoggedIn) {
-      router.push("/login");
-    } else if (props.item.userId !== auth.currentUser.uid) {
-      enqueueSnackbar("You can't edit this card.", { variant: "error" });
-    } else {
-      setEditModalIsOpen(true);
-      handleClose();
-    }
+    setEditModalIsOpen(true);
+    handleClose();
   };
 
   const handleEditModalClose = () => {
@@ -76,19 +49,13 @@ export default function CardMenu(props) {
   };
 
   const handleDeleteClick = async () => {
-    if (!isLoggedIn) {
-      router.push("/login");
-    } else if (props.item.userId !== auth.currentUser.uid) {
-      enqueueSnackbar("You can't delete this card.", { variant: "error" });
-    } else {
-      try {
-        await axiosInstance.delete(`api/posts/${props.item._id}`);
-        setSuccess(true);
-        await mutate("api/posts");
-      } catch (error) {
-        setError(error.message);
-        console.log("error: " + error.message);
-      }
+    try {
+      await axiosInstance.delete(`api/posts/${props.item._id}`);
+      setSuccess(true);
+      await mutate("api/posts");
+    } catch (error) {
+      setError(error.message);
+      console.log("error: " + error.message);
     }
   };
